[BCMOHAD-30696,[BCMOHAD-30510]Resolve merge conflicts
