+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/pages/plr/facility/UpdateFacilityPage.java
@@ -7,6 +7,8 @@ import java.net.URI;
 import java.util.List;
 import java.util.Map;
 
+import ca.bc.gov.health.qa.autotest.plr.fhir.maintain.MaintainOrgBuilder;
+import ca.bc.gov.health.qa.autotest.plr.web.pages.components.AutocompleteMenu;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.components.DropDownMenu;
 import ca.bc.gov.health.qa.autotest.plr.web.tests.model.EndReason;
 
@@ -82,39 +84,6 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 
 	}
 	
-	public void refreshPage() { selenium_.getDriver().navigate().refresh(); }
-
-	public void maxWindow() {
-		selenium_.getDriver().manage().window().maximize();
-	}
-
-
-	
-	private String getSectionContentSelector(FacilitySection section) {
-		return getSectionSelector(section) + "_content";
-	}
-
-	
-	private String getDataBlocksSelector(FacilitySection section) {
-		return getSectionContentSelector(section) + " > table.recordDetailsPanels > tbody > tr > td > div.ui-panel";
-	}
-	
-	public int grabDataBlockCount(FacilitySection section) {
-		return selenium_.findElements(By.cssSelector(getDataBlocksSelector(section))).size();
-	}
-	
-	
-	
-	private String getDataBlockSelector(FacilitySection section, int index) {
-		if (index < 0) {
-			String msg = String.format("Negative index (%d).", index);
-			throw new IllegalArgumentException(msg);
-		}
-		return getDataBlocksSelector(section) + ":nth-of-type(" + (index + 1) + ")";
-	}
-	
-	
-
     /**
      * Gets whether a specified data block is active.
      *
@@ -127,6 +96,13 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 		return selenium_.searchElement(locator) != null;
 	}
 	
+	/**
+	 * get Data Block Header Update Button Selector
+	 *
+	 * @param section
+	 * @param index
+	 * @return CSS selector of Data Block Header Update Button
+	 */
 	protected String getDataBlockHeaderUpdateButtonSelector(FacilitySection section, int index) {
 		return getDataBlockHeaderSelector(section, index) + " > div.ui-panel-actions "
 				+ " > span >a >img[title^='Update']";
@@ -156,13 +132,16 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 		return count;
 	}
 
+	/**
+	 * Using a Javascript Executor to press button
+	 * @param button
+	 */
 	private void clickbuttonWait(WebElement button) {
 		try {
 			button.click();
-		}catch(ElementClickInterceptedException|StaleElementReferenceException e){
-			JavascriptExecutor js = (JavascriptExecutor)selenium_.getDriver();
-		    //js.executeScript("arguments[0].scrollIntoView(true);", button);
-		    js.executeScript("arguments[0].click();", button);
+		} catch (ElementClickInterceptedException | StaleElementReferenceException e) {
+			JavascriptExecutor js = (JavascriptExecutor) selenium_.getDriver();
+			js.executeScript("arguments[0].click();", button);
 		}
 
 	}
@@ -193,6 +172,13 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 
 	}
 	
+	
+	/**
+	 * Get Dialog Css selector
+	 *
+	 * @param section
+	 * @return string of dialog CSS selector
+	 */
 	private String getDialogCss(FacilitySection section) {
 		String dialogName =DIALOG_MAP.get(section).getDialogName();
 		String formName=DIALOG_MAP.get(section).getFormName();
@@ -200,54 +186,103 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 		return "div#"+dialogName+" > div#"+dialogName+"_content" +" > form#"+formName;
 	}
 	
-	private void setEndReasonByVisibleText(FacilitySection section, String visibleText)
-    {
-        String formName=DIALOG_MAP.get(section).getFormName();
-        String endReasonName=DIALOG_MAP.get(section).getEndReasonName();
+	/**
+	 * click Data Block Update Button. Dialog will pop up if successful.
+	 * @param section
+	 * @param visibleText
+	 */
+	private void setEndReasonByVisibleText(FacilitySection section, String visibleText) {
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String endReasonName = DIALOG_MAP.get(section).getEndReasonName();
 
-        DropDownMenu endReasonDrop = new DropDownMenu(selenium_,
-                By.cssSelector("label#" + formName + "\\:" + endReasonName + "_label"),
-                By.cssSelector("div#" + formName + "\\:" + endReasonName + "_panel"));
+		DropDownMenu endReasonDrop = new DropDownMenu(selenium_,
+				By.cssSelector("label#" + formName + "\\:" + endReasonName + "_label"),
+				By.cssSelector("div#" + formName + "\\:" + endReasonName + "_panel"));
 
-        endReasonDrop.selectItem(visibleText);
+		endReasonDrop.selectItem(visibleText);
+	}
+	/**
+	 * set Drop down List By Visible Text
+	 *
+	 * @param section
+	 * @param dropdownName
+	 * @param visibleText
+	 *
+	 */
+	private void setDropdownListByVisibleText(FacilitySection section, String dropdownName, String visibleText) {
+		if (StringUtils.isEmpty(visibleText))
+			return;
+		String formName = DIALOG_MAP.get(section).getFormName();
+
+		DropDownMenu dropdownMenu = new DropDownMenu(selenium_,
+				By.cssSelector("label#" + formName + "\\:" + dropdownName + "_label"),
+				By.cssSelector("div#" + formName + "\\:" + dropdownName + "_panel"));
+
+		dropdownMenu.selectItem(visibleText);
+	}	
+	
+	/**
+	 * select Drop down List By DropMenu
+	 *
+	 * @param section
+	 * @param dropDownName
+	 * @param dropdownText
+	 */
+	private void selectDropdownListByDropMenu(FacilitySection section, String dropDownName, String dropdownText) {
+
+		String formName = DIALOG_MAP.get(section).getFormName();
+
+		DropDownMenu dropdownMenu = new DropDownMenu(selenium_,
+				By.cssSelector("label#" + formName + "\\:" + dropDownName + "_label"),
+				By.cssSelector("div#" + formName + "\\:" + dropDownName + "_panel"));
+
+		dropdownMenu.selectItem(dropdownText);
+
 	}
 	
-	
-	private void setDropdownListByVisibleText(FacilitySection section, String dropdownName,String visibleText)
-    {
-		String formName=DIALOG_MAP.get(section).getFormName();
-
-        DropDownMenu dropdownMenu = new DropDownMenu(selenium_,
-                By.cssSelector("label#" + formName + "\\:" + dropdownName + "_label"),
-                By.cssSelector("div#" + formName + "\\:" + dropdownName + "_panel"));
-
-        dropdownMenu.selectItem(visibleText);
-	}
-	
+	/**
+	 * Click Dialog Submit Button
+	 *
+	 * @param section
+	 *
+	 */
 	public void clickDialogSubmitButton(FacilitySection section) {
 		clickDialogSubmitButton(section, false);
 	}
 
-	private void clickDialogSubmitButton(FacilitySection section,boolean expectError) {
-		
-		String formName=DIALOG_MAP.get(section).getFormName();
-		String submitButtonName=DIALOG_MAP.get(section).getSubmitButtonName();
-		String dialogCss=getDialogCss(section);
-		
-		String buttonCss=dialogCss+" > div.formControls"+" > button#"+formName+"\\:"+submitButtonName;
-		WebElement button=selenium_.findElement(By.cssSelector(buttonCss));
+	/**
+	 * Click the "Update/Add" dialog submission button. If an error is anticipated,
+	 * wait for the error message to appear.
+	 *
+	 * @param section
+	 *
+	 */
+	private void clickDialogSubmitButton(FacilitySection section, boolean expectError) {
+
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String submitButtonName = DIALOG_MAP.get(section).getSubmitButtonName();
+		String dialogCss = getDialogCss(section);
+
+		String buttonCss = dialogCss + " > div.formControls" + " > button#" + formName + "\\:" + submitButtonName;
+		WebElement button = selenium_.findElement(By.cssSelector(buttonCss));
 		button.click();
 		waitSeconds(2);
-		if(expectError) {
-			//selenium_.waitUntil(ExpectedConditions.visibilityOf(button));
+		if (expectError) {
 			selenium_.waitUntil(SeleniumExpectedConditions.pageToBeReady());
 		}
 	}
 	
+	/**
+	 * wait Error Message showing up, and return a copy of message as result
+	 * note the result is a set of messages, if there are more than one error messages
+	 *
+	 * @param section
+	 * @return String of error messages
+	 */
 	public String waitErrorMessage(FacilitySection section) {
 
 		String msgDisplay = "";
-		
+
 		msgDisplay = getDialogMessages(section);
 		while (StringUtils.isEmpty(msgDisplay)) {
 			waitSeconds(5);
@@ -265,73 +300,103 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 
 	}
 	
-	
+	/**
+	 * Click the Header Add Date Block button. A dialog box will appear upon successful completion.
+	 *
+	 * @param section
+	 *
+	 */
+	public void clickHeaderAddDateBlockButton(FacilitySection section) {
+		String title = DIALOG_MAP.get(section).getAddButtonImgText();
+		String clickElementCss = getSectionSelector(section) + " > div > div >a > img[title='" + title + "'";
 
-	private void clickHeaderAddDateBlockButton(FacilitySection section) {
-		String title=DIALOG_MAP.get(section).getAddButtonImgText();
-		String clickElementCss=getSectionSelector(section)+" > div > div >a > img[title='"+title+"'";
-		//WebElement clickElement=selenium_.findElement(By.cssSelector(clickElementCss));
-		WebElement clickElement =selenium_.waitUntil(ExpectedConditions.elementToBeClickable(By.cssSelector(clickElementCss)));
+		WebElement clickElement = selenium_
+				.waitUntil(ExpectedConditions.elementToBeClickable(By.cssSelector(clickElementCss)));
 		selenium_.scrollIntoView(clickElement);
 		try {
 			clickElement.click();
-	    } catch (StaleElementReferenceException | ElementClickInterceptedException e) {
-	        // Re-locate the element and retry the action
-	    	waitSeconds(2);
-	    	clickElement =selenium_.findElement(By.cssSelector(clickElementCss));
-	        clickElement.click();
-	    }
-	
-		
-		String dialogCss=getDialogCss(section);
-		WebElement visibleElement = selenium_.waitUntil(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(dialogCss)));
-		//waitSeconds(2);
+		} catch (StaleElementReferenceException | ElementClickInterceptedException e) {
+			waitSeconds(2);
+			clickElement = selenium_.findElement(By.cssSelector(clickElementCss));
+			clickElement.click();
+		}
+
+		String dialogCss = getDialogCss(section);
+		selenium_
+				.waitUntil(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(dialogCss)));
 	}
 	
-	private void clickHeaderAddDateBlockButtonOrgID(FacilitySection section) {
-		String title=DIALOG_MAP.get(section).getAddButtonImgText();
-		String clickElementCss=getSectionSelector(section)+" > div > div >a > img[title='"+title+"'";
-		//WebElement clickElement=selenium_.findElement(By.cssSelector(clickElementCss));
-		WebElement clickElement =selenium_.waitUntil(ExpectedConditions.elementToBeClickable(By.cssSelector(clickElementCss)));
-		selenium_.scrollIntoView(clickElement);
+	/**
+	 * check if Header Add Date Block Button Displayed
+	 *
+	 * @param section
+	 * @return True if "Add Date Block" Button Displayed, otherwise return false
+	 */
+	public boolean isHeaderAddDateBlockButtonDisplayed(FacilitySection section) {
+		String title = DIALOG_MAP.get(section).getAddButtonImgText();
+		String clickElementCss = getSectionSelector(section) + " > div > div >a > img[title='" + title + "'";
+		WebElement updateButton = null;
 
-		clickbuttonWait(clickElement);
+		try {
+			updateButton = selenium_.findElement(By.cssSelector(clickElementCss));
 
-		String dialogCss=getDialogCss(section);
-		WebElement visibleElement = selenium_.waitUntil(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(dialogCss)));//TO DO
+		} catch (org.openqa.selenium.NoSuchElementException e) {
+			return false;
+		}
+		selenium_.scrollIntoView(updateButton);
+		return updateButton.isDisplayed();
 
 	}
+	
+	
 
-	private void setDialogEffectiveFromAndEffectiveTo(FacilitySection section,String effectiveFrom, String effectiveTo) {
-		String dialogCss=getDialogCss(section);
-		String formName=DIALOG_MAP.get(section).getFormName();
-		String effectiveFromStr=DIALOG_MAP.get(section).getEffectiveFromStr();
-		String effectiveToStr=DIALOG_MAP.get(section).getEffectiveToStr();
-		//<span id="maintainFacilityNameForm:effectiveStartDate"class="ui-calendar ui-trigger-calendar"><input	id="maintainFacilityNameForm:effectiveStartDate_input"
-		String effectFromCss=dialogCss+" >span#"+formName+"\\:"+effectiveFromStr+" >input#"+formName+"\\:"+effectiveFromStr+"_input";
-		WebElement effectFromElement=selenium_.findElement(By.cssSelector(effectFromCss));
+	/**
+	 * set Dialog Effective From And Effective To.
+	 *
+	 * @param section
+	 * @param effectiveFrom
+	 * @param effectiveTo
+	 */
+	private void setDialogEffectiveFromAndEffectiveTo(FacilitySection section, String effectiveFrom,
+			String effectiveTo) {
+		String dialogCss = getDialogCss(section);
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String effectiveFromStr = DIALOG_MAP.get(section).getEffectiveFromStr();
+		String effectiveToStr = DIALOG_MAP.get(section).getEffectiveToStr();
+
+		String effectFromCss = dialogCss + " >span#" + formName + "\\:" + effectiveFromStr + " >input#" + formName
+				+ "\\:" + effectiveFromStr + "_input";
+		WebElement effectFromElement = selenium_.findElement(By.cssSelector(effectFromCss));
 		effectFromElement.clear();
-		if(!StringUtils.isEmpty(effectiveFrom))effectFromElement.sendKeys(effectiveFrom);
-		
-		//<span id="maintainFacilityNameForm:effectiveEndDate"class="ui-calendar ui-trigger-calendar"><inputid="maintainFacilityNameForm:effectiveEndDate_input"
-		String effectToCss=dialogCss+" >span#"+formName+"\\:"+effectiveToStr+" >input#"+formName+"\\:"+effectiveToStr+"_input";
-		WebElement effectToElement=selenium_.findElement(By.cssSelector(effectToCss));
-		effectToElement.clear();
-		if(!StringUtils.isEmpty(effectiveTo))effectToElement.sendKeys(effectiveTo);				
-	}
+		if (!StringUtils.isEmpty(effectiveFrom))
+			effectFromElement.sendKeys(effectiveFrom);
 
+		String effectToCss = dialogCss + " >span#" + formName + "\\:" + effectiveToStr + " >input#" + formName + "\\:"
+				+ effectiveToStr + "_input";
+		WebElement effectToElement = selenium_.findElement(By.cssSelector(effectToCss));
+		effectToElement.clear();
+		if (!StringUtils.isEmpty(effectiveTo))
+			effectToElement.sendKeys(effectiveTo);
+	}
+	
+
+	/**
+	 * Click Data Bloc kUpdate Button. Dialog will pop up if successful.
+	 * @param section
+	 * @return String of Dialog Messages
+	 */
 	private String getDialogMessages(FacilitySection section) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(section).getFormName();
-		String dialogCss=getDialogCss(section);
-		String alertMsgCss=dialogCss+" >div#"+formName+"\\:"+"messages";
-		WebElement alertMsg=selenium_.findElement(By.cssSelector(alertMsgCss));
-		if(alertMsg.isDisplayed()) {
-			msgDisplay=alertMsg.getText();
-			
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String dialogCss = getDialogCss(section);
+		String alertMsgCss = dialogCss + " >div#" + formName + "\\:" + "messages";
+		WebElement alertMsg = selenium_.findElement(By.cssSelector(alertMsgCss));
+		if (alertMsg.isDisplayed()) {
+			msgDisplay = alertMsg.getText();
+
 		}
 		return msgDisplay;
-		
+
 	}
 	
 	private void selectDropdownListByClick(FacilitySection section, String dropDownName, String dropdownText) {
@@ -376,50 +441,119 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 				
 		
 	}
+	
+	/**
+	 * Verify if a DropdownList is displayed in a Update Dialog
+	 *
+	 * @param section
+	 * @param index
+	 * @param dropDownName
+	 * @return True if the DropdownList is displayed, otherwise false
+	 */
+	public boolean findUpdateDialogDropdownList(FacilitySection section, int index, String dropDownName) {
 
-    /**
-     * TODO (KD) - doc
-     *
-     * @param section
-     * @param index
-     */
-	public void ceaseDataBlock(FacilitySection section,int index) {
-		
-		String formName=DIALOG_MAP.get(section).getFormName();
-		String submitButtonName=DIALOG_MAP.get(section).getSubmitButtonName();
-		
+		clickDataBlockUpdateButton(section, index);
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String dialogCss = getDialogCss(section);
+		String dropListCss = dialogCss + " > div#" + formName + "\\:" + dropDownName
+				+ " >div.ui-helper-hidden-accessible" + " > select#" + formName + "\\:" + dropDownName + "_input";
+		boolean find = true;
+		try {
+			WebElement dropList = selenium_.findElement(By.cssSelector(dropListCss));
+		} catch (Exception e) {
+			find = false;
+		}
+
+		WebElement cancelButton = selenium_.findElement(By.linkText("Cancel"));
+		selenium_.scrollIntoView(cancelButton);
+		cancelButton.click();
+
+		return find;
+
+	}
+	
+	/**
+	 * Verify if a input element is displayed in a Update Dialog
+	 *
+	 * @param section
+	 * @param index
+	 * @param inputName
+	 *
+	 * @return true if the input element displayed, otherwise return false
+	 */
+	public boolean findUpdateDialogInput(FacilitySection section, int index, String inputName) {
+
+		clickDataBlockUpdateButton(section, index);
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String dialogCss = getDialogCss(section);
+		String inputIdCss = dialogCss + " >input#" + formName + "\\:" + inputName;
+		boolean find = true;
+
+		try {
+			WebElement inputId = selenium_.findElement(By.cssSelector(inputIdCss));
+		} catch (Exception e) {
+			find = false;
+		}
+		WebElement cancelButton = selenium_.findElement(By.linkText("Cancel"));
+		selenium_.scrollIntoView(cancelButton);
+		cancelButton.click();
+
+		return find;
+
+	}
+	
+
+	/**
+	 * Cease Data Block
+	 *
+	 * @param section
+	 * @param index
+	 */
+	public void ceaseDataBlock(FacilitySection section, int index) {
+
+		String formName = DIALOG_MAP.get(section).getFormName();
+		String submitButtonName = DIALOG_MAP.get(section).getSubmitButtonName();
+
 		clickDataBlockUpdateButton(section, index);
 		waitSeconds(2);
-		
-		String dialogCss=getDialogCss(section);
-			
-		setEndReasonByVisibleText(section,EndReason.CEASE.getText());
-		
-		String buttonCss=dialogCss+" > div.formControls"+" > button#"+formName+"\\:"+submitButtonName;
-		WebElement button=selenium_.findElement(By.cssSelector(buttonCss));
+
+		String dialogCss = getDialogCss(section);
+
+		setEndReasonByVisibleText(section, EndReason.CEASE.getText());
+
+		String buttonCss = dialogCss + " > div.formControls" + " > button#" + formName + "\\:" + submitButtonName;
+		WebElement button = selenium_.findElement(By.cssSelector(buttonCss));
 		button.click();
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 	}
-
-    /**
-     * TODO (KD) - doc
-     *
-     * @param section
-     */
+	
+	/**
+	 * Cease All Data Block Under Section
+	 *
+	 * @param section
+	 */
 	public void ceaseAllDataBlockUnderSection(FacilitySection section) {
-		int count=grabActiveDataBlockCount(section, true);
-		for(int i=0;i<count;i++) { 
-			ceaseDataBlock(section,0);
+		int count = grabActiveDataBlockCount(section, true);
+		for (int i = 0; i < count; i++) {
+			ceaseDataBlock(section, 0);
 		}
 	}
 
-    private void findAndFillField(String dialogCss, String formName, String field, String fieldCss)
-    {
-        String inputNameCss=dialogCss+" >input#"+formName+"\\:"+fieldCss;
-        WebElement inputName=selenium_.findElement(By.cssSelector(inputNameCss));
-        inputName.clear();
-        if(!StringUtils.isEmpty(field))inputName.sendKeys(field);
-    }
+	/**
+	 * Find And Fill an input Field
+	 *
+	 * @param dialogCss
+	 * @param formName
+	 * @param field
+	 * @param fieldCss
+	 */
+	private void findAndFillField(String dialogCss, String formName, String field, String fieldCss) {
+		String inputNameCss = dialogCss + " >input#" + formName + "\\:" + fieldCss;
+		WebElement inputName = selenium_.findElement(By.cssSelector(inputNameCss));
+		inputName.clear();
+		if (!StringUtils.isEmpty(field))
+			inputName.sendKeys(field);
+	}
 
     /**
      * Attempts to add a name data block with provided values
@@ -431,31 +565,27 @@ public class UpdateFacilityPage extends ViewFacilityPage {
      * @return					the full message dialog of errors, if any exist. otherwise an empty string
      */
 	public String addNameDataBlock(String name, String desc, String effectiveFrom, String effectiveTo) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.NAMES).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.NAMES);
-		
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.NAMES).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.NAMES);
+
 		clickHeaderAddDateBlockButton(FacilitySection.NAMES);
-		//waitSeconds(2);
-		
-		//<input id="maintainFacilityNameForm:shortName"
-        findAndFillField(dialogCss,formName,name,"shortName");
-		
-		//<input id="maintainFacilityNameForm:description"
-        findAndFillField(dialogCss,formName,desc,"description");
-		
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NAMES,effectiveFrom,effectiveTo);
-		
+
+		findAndFillField(dialogCss, formName, name, "shortName");
+
+		findAndFillField(dialogCss, formName, desc, "description");
+
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NAMES, effectiveFrom, effectiveTo);
+
 		clickDialogSubmitButton(FacilitySection.NAMES);
-		
-		msgDisplay=getDialogMessages(FacilitySection.NAMES);
-		
-		if(!StringUtils.isEmpty(msgDisplay)){
-			//cancel button
-			WebElement cancelButton=selenium_.findElement(By.linkText("Cancel"));
+
+		msgDisplay = getDialogMessages(FacilitySection.NAMES);
+
+		if (!StringUtils.isEmpty(msgDisplay)) {
+			WebElement cancelButton = selenium_.findElement(By.linkText("Cancel"));
 			cancelButton.click();
 		}
-		
+
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 		return msgDisplay;
 	}
@@ -471,32 +601,29 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 	 * @param endReason			the end reason to specify when updating the data block
      * @return					a string of the full message dialog of errors, if any exist. otherwise an empty string
      */
-	public String updateNameDataBlock(String name, String desc, String effectiveFrom, String effectiveTo, int index, String endReason) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.NAMES).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.NAMES);
-		
+	public String updateNameDataBlock(String name, String desc, String effectiveFrom, String effectiveTo,
+			EndReason endReasonCode, int index) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.NAMES).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.NAMES);
+
 		clickDataBlockUpdateButton(FacilitySection.NAMES, index);
-		//waitSeconds(2);
-		
-		//<input id="maintainFacilityNameForm:shortName"
-        findAndFillField(dialogCss,formName,name,"shortName");
-		
-		//<input id="maintainFacilityNameForm:description"
-        findAndFillField(dialogCss,formName,desc,"description");
-		
-		setEndReasonByVisibleText(FacilitySection.NAMES,endReason);
-		
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NAMES,effectiveFrom,effectiveTo);
-		
-		
+
+		findAndFillField(dialogCss, formName, name, "shortName");
+
+		findAndFillField(dialogCss, formName, desc, "description");
+		if (endReasonCode != null)
+			setEndReasonByVisibleText(FacilitySection.NAMES, endReasonCode.getText());
+
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NAMES, effectiveFrom, effectiveTo);
+
 		clickDialogSubmitButton(FacilitySection.NAMES);
-		
-		msgDisplay=getDialogMessages(FacilitySection.NAMES);
-		
-		if(!StringUtils.isEmpty(msgDisplay)){
-			//cancel button
-			WebElement cancelButton=selenium_.findElement(By.linkText("Cancel"));
+
+		msgDisplay = getDialogMessages(FacilitySection.NAMES);
+
+		if (!StringUtils.isEmpty(msgDisplay)) {
+			// cancel button
+			WebElement cancelButton = selenium_.findElement(By.linkText("Cancel"));
 			cancelButton.click();
 		}
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
@@ -541,45 +668,44 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 		return msgDisplay;
 	}
 
-    /**
-     * TODO (KD) - doc
-     *
-     * @param id
-     * @param text
-     * @param effectiveFrom
-     * @param effectiveTo
-     * @param expectError
-     * @return
-     */
-	public String addNoteDataBlock(String id, String text, String effectiveFrom,
-			String effectiveTo,boolean expectError) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.NOTES).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.NOTES);
-		
+	/**
+	 * Add Note Data Block
+	 *
+	 * @param id
+	 * @param text
+	 * @param effectiveFrom
+	 * @param effectiveTo
+	 * @param expectError
+	 * @return If a predicted exception occurs, it will capture and return the
+	 *         associated error message; otherwise, it waits for the dialog to
+	 *         dismiss before returning an empty string.
+	 */
+	public String addNoteDataBlock(String id, String text, String effectiveFrom, String effectiveTo,
+			boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.NOTES).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.NOTES);
+
 		clickHeaderAddDateBlockButton(FacilitySection.NOTES);
-		//waitSeconds(2);
-		
-		
-		//selectDropdownListByClick(FacilitySection.IDENTIFIERS,"identifierType",idType);
-		
-		String inputIdCss=dialogCss+" >input#"+formName+"\\:"+"identifier";
-		WebElement inputId=selenium_.findElement(By.cssSelector(inputIdCss));
+		String inputIdCss = dialogCss + " >input#" + formName + "\\:" + "identifier";
+		WebElement inputId = selenium_.findElement(By.cssSelector(inputIdCss));
 		inputId.clear();
-		if(!StringUtils.isEmpty(id))inputId.sendKeys(id);
-		
-		String inputTextCss=dialogCss+" >textarea#"+formName+"\\:"+"noteText";
-		WebElement inputText=selenium_.findElement(By.cssSelector(inputTextCss));
+		if (!StringUtils.isEmpty(id))
+			inputId.sendKeys(id);
+
+		String inputTextCss = dialogCss + " >textarea#" + formName + "\\:" + "noteText";
+		WebElement inputText = selenium_.findElement(By.cssSelector(inputTextCss));
 		inputText.clear();
-		if(!StringUtils.isEmpty(text))inputText.sendKeys(text);
-		
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NOTES,effectiveFrom,effectiveTo);
-		
+		if (!StringUtils.isEmpty(text))
+			inputText.sendKeys(text);
+
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NOTES, effectiveFrom, effectiveTo);
+
 		clickDialogSubmitButton(FacilitySection.NOTES);
-		
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.NOTES);
-		
+
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.NOTES);
+
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 
 		return msgDisplay;
@@ -596,102 +722,102 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 	 * @param endReason			the end reason field to fill the end reason code field with
      * @return					a string of the error message, if expectError is true. otherwise an empty string
      */
-	public String updateNoteDataBlock( String text, String effectiveFrom, String effectiveTo,
-			int index, boolean expectError, String endReason) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.NOTES).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.NOTES);
-		
+	public String updateNoteDataBlock(String text, String effectiveFrom, String effectiveTo, EndReason endReasonCode,
+			int index, boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.NOTES).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.NOTES);
+
 		clickDataBlockUpdateButton(FacilitySection.NOTES, index);
-		//waitSeconds(2);
-		
-		String inputTextCss=dialogCss+" >textarea#"+formName+"\\:"+"noteText";
-		WebElement inputText=selenium_.findElement(By.cssSelector(inputTextCss));
+		waitSeconds(2);
+
+		String inputTextCss = dialogCss + " >textarea#" + formName + "\\:" + "noteText";
+		WebElement inputText = selenium_.findElement(By.cssSelector(inputTextCss));
 		inputText.clear();
-		if(!StringUtils.isEmpty(text))inputText.sendKeys(text);
-		
-		
-		setEndReasonByVisibleText(FacilitySection.NOTES,endReason);
-		
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NOTES,effectiveFrom,effectiveTo);
-		
-		
+		if (!StringUtils.isEmpty(text))
+			inputText.sendKeys(text);
+
+		if (endReasonCode != null)
+			setEndReasonByVisibleText(FacilitySection.NOTES, endReasonCode.getText());
+
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.NOTES, effectiveFrom, effectiveTo);
+
 		clickDialogSubmitButton(FacilitySection.NOTES);
-		
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.NOTES);
-				
+
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.NOTES);
+
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
-		return msgDisplay;	
+		return msgDisplay;
 	}
 
-    /**
-     * TODO (KD) - doc
-     *
-     * @param idType
-     * @param id
-     * @param relationType
-     * @param effectiveFrom
-     * @param effectiveTo
-     * @param expectError
-     * @return
-     */
-	public String addRelatedOrganizationDataBlock(String idType, String id, String relationType,String effectiveFrom,
+
+	/**
+	 * Add Related Organization Data Block
+	 *
+	 * @param idType
+	 * @param id
+	 * @param relationType
+	 * @param effectiveFrom
+	 * @param effectiveTo
+	 * @param expectError
+	 * @return If a predicted exception occurs, it will capture and return the
+	 *         associated error message; otherwise, it waits for the dialog to
+	 *         dismiss before returning an empty string.
+	 */
+	public String addRelatedOrganizationDataBlock(String idType, String id, String relationType, String effectiveFrom,
 			String effectiveTo, boolean expectError) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.ORGANIZATION_RELATIONSHIPS).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.ORGANIZATION_RELATIONSHIPS);
-		
-		clickHeaderAddDateBlockButtonOrgID(FacilitySection.ORGANIZATION_RELATIONSHIPS);
-		//waitSeconds(2);
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.ORGANIZATION_RELATIONSHIPS).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.ORGANIZATION_RELATIONSHIPS);
 
+		clickHeaderAddDateBlockButton(FacilitySection.ORGANIZATION_RELATIONSHIPS);
 
-		selectDropdownListByClick(FacilitySection.ORGANIZATION_RELATIONSHIPS,"providerType",idType);
-		
-		String inputIdCss=dialogCss+" >input#"+formName+"\\:"+"rpi";
-		WebElement inputId=selenium_.findElement(By.cssSelector(inputIdCss));
+		setDropdownListByVisibleText(FacilitySection.ORGANIZATION_RELATIONSHIPS, "providerType", idType);
+
+		String inputIdCss = dialogCss + " >input#" + formName + "\\:" + "rpi";
+		WebElement inputId = selenium_.findElement(By.cssSelector(inputIdCss));
 		inputId.clear();
-		if(!StringUtils.isEmpty(id))inputId.sendKeys(id);
-		
-		selectDropdownListByClick(FacilitySection.ORGANIZATION_RELATIONSHIPS,"relationshipType",relationType);
-		
-		if(expectError)waitSeconds(2);
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ORGANIZATION_RELATIONSHIPS,effectiveFrom,effectiveTo);
-		
-		
-		clickDialogSubmitButton(FacilitySection.ORGANIZATION_RELATIONSHIPS,expectError);
-		
-		//msgDisplay=getDialogMessages(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+		if (!StringUtils.isEmpty(id))
+			inputId.sendKeys(id);
+
+		setDropdownListByVisibleText(FacilitySection.ORGANIZATION_RELATIONSHIPS, "relationshipType", relationType);
+
+		if (expectError)
+			waitSeconds(2);
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ORGANIZATION_RELATIONSHIPS, effectiveFrom, effectiveTo);
+
+		clickDialogSubmitButton(FacilitySection.ORGANIZATION_RELATIONSHIPS, expectError);
+
 		if (expectError) {
 			waitSeconds(10);
-			
+
 			msgDisplay = getDialogMessages(FacilitySection.ORGANIZATION_RELATIONSHIPS);
-			while (StringUtils.isEmpty(msgDisplay) ) {
+			while (StringUtils.isEmpty(msgDisplay)) {
 				waitSeconds(5);
 				try {
-				msgDisplay = getDialogMessages(FacilitySection.ORGANIZATION_RELATIONSHIPS);}
-				catch(StaleElementReferenceException  e) {
+					msgDisplay = getDialogMessages(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+				} catch (StaleElementReferenceException e) {
 					waitSeconds(5);
 				}
 			}
-			WebElement cancelButton=selenium_.findElement(By.linkText("Cancel"));
+			WebElement cancelButton = selenium_.findElement(By.linkText("Cancel"));
 			selenium_.scrollIntoView(cancelButton);
 			cancelButton.click();
 		}
-		
+
 		try {
-		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));}
-		catch(org.openqa.selenium.TimeoutException e) {
-			msgDisplay=getDialogMessages(FacilitySection.ORGANIZATION_RELATIONSHIPS);
-			WebElement cancelButton=selenium_.findElement(By.linkText("Cancel"));
+			selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
+		} catch (org.openqa.selenium.TimeoutException e) {
+			msgDisplay = getDialogMessages(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+			WebElement cancelButton = selenium_.findElement(By.linkText("Cancel"));
 			selenium_.scrollIntoView(cancelButton);
 			cancelButton.click();
 		}
 		return msgDisplay;
 	}
-
 	/**
-	 * TODO (AMV) - doc
+	 * update RelationshipOrg Data Block
 	 *
 	 * @param effectiveFrom
 	 * @param effectiveTo
@@ -700,27 +826,28 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 	 * @param endReason
 	 * @return
 	 */
-	public String updateRelationshipOrgBlock(String effectiveFrom, String effectiveTo, int index, boolean expectError, String endReason)
-	{
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.ORGANIZATION_RELATIONSHIPS).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+	public String updateRelatedOrganizationDataBlock(String effectiveFrom, String effectiveTo, EndReason endReason,
+			int index, boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.ORGANIZATION_RELATIONSHIPS).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.ORGANIZATION_RELATIONSHIPS);
 
 		clickDataBlockUpdateButton(FacilitySection.ORGANIZATION_RELATIONSHIPS, index);
-		//waitSeconds(2);
 
-		setEndReasonByVisibleText(FacilitySection.ORGANIZATION_RELATIONSHIPS,endReason);
+		if (endReason != null)
+			setEndReasonByVisibleText(FacilitySection.ORGANIZATION_RELATIONSHIPS, endReason.getText());
 
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ORGANIZATION_RELATIONSHIPS,effectiveFrom,effectiveTo);
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ORGANIZATION_RELATIONSHIPS, effectiveFrom, effectiveTo);
 
 		clickDialogSubmitButton(FacilitySection.ORGANIZATION_RELATIONSHIPS);
 
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.ORGANIZATION_RELATIONSHIPS);
 
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 		return msgDisplay;
 	}
+
 
     /**
      * Attempts to add a telecommunication data block with provided values
@@ -734,44 +861,45 @@ public class UpdateFacilityPage extends ViewFacilityPage {
      * @param expectError		whether an error is anticipated (true) or not (false)
      * @return					a string of the error message, if expectError is true. otherwise an empty string
      */
-	public String addTelecommunicationDataBlock(String type, String areaCode, String phoneNumber,String extension,
-			String effectiveFrom,String effectiveTo, boolean expectError) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.TELECOMMUNICATIONS).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.TELECOMMUNICATIONS);
-		
+	public String addTelecommunicationDataBlock(String telecomType, String areaCode, String phoneNumber, String extension,
+			String effectiveFrom, String effectiveTo, boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.TELECOMMUNICATIONS).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.TELECOMMUNICATIONS);
+
 		clickHeaderAddDateBlockButton(FacilitySection.TELECOMMUNICATIONS);
-		//waitSeconds(2);
-		
-		
-		selectDropdownListByClick(FacilitySection.TELECOMMUNICATIONS,"telecomType",type);
-		
-		String inputAreaCodeCss=dialogCss+" >input#"+formName+"\\:"+"AreaCode";
-		WebElement inputAreaCode=selenium_.findElement(By.cssSelector(inputAreaCodeCss));
+
+		setDropdownListByVisibleText(FacilitySection.TELECOMMUNICATIONS, "telecomType", telecomType);
+
+		String inputAreaCodeCss = dialogCss + " >input#" + formName + "\\:" + "AreaCode";
+		WebElement inputAreaCode = selenium_.findElement(By.cssSelector(inputAreaCodeCss));
 		inputAreaCode.clear();
-		if(!StringUtils.isEmpty(areaCode))inputAreaCode.sendKeys(areaCode);
-		
-		String inputPhoneNumberCss=dialogCss+" >input#"+formName+"\\:"+"Phone_Number";
-		WebElement inputPhoneNumber=selenium_.findElement(By.cssSelector(inputPhoneNumberCss));
+		if (!StringUtils.isEmpty(areaCode))
+			inputAreaCode.sendKeys(areaCode);
+
+		String inputPhoneNumberCss = dialogCss + " >input#" + formName + "\\:" + "Phone_Number";
+		WebElement inputPhoneNumber = selenium_.findElement(By.cssSelector(inputPhoneNumberCss));
 		inputPhoneNumber.clear();
-		if(!StringUtils.isEmpty(phoneNumber))inputPhoneNumber.sendKeys(phoneNumber);
-		
-		String inputExtensionCss=dialogCss+" >input#"+formName+"\\:"+"extension";
-		WebElement inputExtension=selenium_.findElement(By.cssSelector(inputExtensionCss));
+		if (!StringUtils.isEmpty(phoneNumber))
+			inputPhoneNumber.sendKeys(phoneNumber);
+
+		String inputExtensionCss = dialogCss + " >input#" + formName + "\\:" + "extension";
+		WebElement inputExtension = selenium_.findElement(By.cssSelector(inputExtensionCss));
 		inputExtension.clear();
-		if(!StringUtils.isEmpty(extension))inputExtension.sendKeys(extension);
-		
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.TELECOMMUNICATIONS,effectiveFrom,effectiveTo);
-		
-		
+		if (!StringUtils.isEmpty(extension))
+			inputExtension.sendKeys(extension);
+
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.TELECOMMUNICATIONS, effectiveFrom, effectiveTo);
+
 		clickDialogSubmitButton(FacilitySection.TELECOMMUNICATIONS);
-		
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.TELECOMMUNICATIONS);
-		
+
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.TELECOMMUNICATIONS);
+
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 		return msgDisplay;
 	}
+
 
 	/**
 	 * Attempts to update a telecommunication data block with provided values
@@ -786,42 +914,47 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 	 * @param endReason			the end reason to fill the end reason code field with
 	 * @return					a string of the error message, if expectError is true. otherwise an empty string
 	 */
-	public String updateTelecommunicationBlock(String areaCode, String phoneNumber, String extension,
-											   String effectiveFrom, String effectiveTo, int index, boolean expectError, String endReason)
-	{
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.TELECOMMUNICATIONS).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.TELECOMMUNICATIONS);
+
+	public String updateTelecommunicationDataBlock(String areaCode, String phoneNumber, String extension,
+			String effectiveFrom, String effectiveTo, EndReason endReasonCode, int index, boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.TELECOMMUNICATIONS).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.TELECOMMUNICATIONS);
 
 		clickDataBlockUpdateButton(FacilitySection.TELECOMMUNICATIONS, index);
-		//waitSeconds(2);
+		waitSeconds(2);
 
-		String inputAreaCodeCss=dialogCss+" >input#"+formName+"\\:"+"AreaCode";
-		WebElement inputAreaCode=selenium_.findElement(By.cssSelector(inputAreaCodeCss));
+		String inputAreaCodeCss = dialogCss + " >input#" + formName + "\\:" + "AreaCode";
+		WebElement inputAreaCode = selenium_.findElement(By.cssSelector(inputAreaCodeCss));
 		inputAreaCode.clear();
-		if(!StringUtils.isEmpty(areaCode))inputAreaCode.sendKeys(areaCode);
+		if (!StringUtils.isEmpty(areaCode))
+			inputAreaCode.sendKeys(areaCode);
 
-		String inputPhoneNumberCss=dialogCss+" >input#"+formName+"\\:"+"Phone_Number";
-		WebElement inputPhoneNumber=selenium_.findElement(By.cssSelector(inputPhoneNumberCss));
+		String inputPhoneNumberCss = dialogCss + " >input#" + formName + "\\:" + "Phone_Number";
+		WebElement inputPhoneNumber = selenium_.findElement(By.cssSelector(inputPhoneNumberCss));
 		inputPhoneNumber.clear();
-		if(!StringUtils.isEmpty(phoneNumber))inputPhoneNumber.sendKeys(phoneNumber);
+		if (!StringUtils.isEmpty(phoneNumber))
+			inputPhoneNumber.sendKeys(phoneNumber);
 
-		String inputExtensionCss=dialogCss+" >input#"+formName+"\\:"+"extension";
-		WebElement inputExtension=selenium_.findElement(By.cssSelector(inputExtensionCss));
+		String inputExtensionCss = dialogCss + " >input#" + formName + "\\:" + "extension";
+		WebElement inputExtension = selenium_.findElement(By.cssSelector(inputExtensionCss));
 		inputExtension.clear();
-		if(!StringUtils.isEmpty(extension))inputExtension.sendKeys(extension);
+		if (!StringUtils.isEmpty(extension))
+			inputExtension.sendKeys(extension);
 
-		setEndReasonByVisibleText(FacilitySection.TELECOMMUNICATIONS, endReason);
+		if (endReasonCode != null)
+			setEndReasonByVisibleText(FacilitySection.TELECOMMUNICATIONS, endReasonCode.getText());
 
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.TELECOMMUNICATIONS,effectiveFrom,effectiveTo);
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.TELECOMMUNICATIONS, effectiveFrom, effectiveTo);
 
 		clickDialogSubmitButton(FacilitySection.TELECOMMUNICATIONS);
 
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.TELECOMMUNICATIONS);
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.TELECOMMUNICATIONS);
 
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 		return msgDisplay;
+
 	}
 
     /**
@@ -834,34 +967,29 @@ public class UpdateFacilityPage extends ViewFacilityPage {
      * @param expectError		whether an error is anticipated (true) or not (false)
      * @return					a string of the error message, if expectError is true. otherwise an empty string
      */
-	public String addElectronicAddressDataBlock(String type, String address,
-			String effectiveFrom,String effectiveTo,boolean expectError) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.ELECTRONIC_ADDRESSES).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.ELECTRONIC_ADDRESSES);
-		
+	public String addElectronicAddressDataBlock(String type, String address, String effectiveFrom, String effectiveTo,
+			boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.ELECTRONIC_ADDRESSES).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.ELECTRONIC_ADDRESSES);
+
 		clickHeaderAddDateBlockButton(FacilitySection.ELECTRONIC_ADDRESSES);
-		//waitSeconds(2);
-		
-		
-		selectDropdownListByClick(FacilitySection.ELECTRONIC_ADDRESSES,"type",type);
-		
-		String inputAddressCss=dialogCss+" >input#"+formName+"\\:"+"electronicAddress";
-		WebElement inputAddress=selenium_.findElement(By.cssSelector(inputAddressCss));
+
+		setDropdownListByVisibleText(FacilitySection.ELECTRONIC_ADDRESSES, "type", type);
+
+		String inputAddressCss = dialogCss + " >input#" + formName + "\\:" + "electronicAddress";
+		WebElement inputAddress = selenium_.findElement(By.cssSelector(inputAddressCss));
 		inputAddress.clear();
-		if(!StringUtils.isEmpty(address))inputAddress.sendKeys(address);
+		if (!StringUtils.isEmpty(address))
+			inputAddress.sendKeys(address);
 
-		
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ELECTRONIC_ADDRESSES,effectiveFrom,effectiveTo);
-		
-		
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ELECTRONIC_ADDRESSES, effectiveFrom, effectiveTo);
+
 		clickDialogSubmitButton(FacilitySection.ELECTRONIC_ADDRESSES);
-		
 
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.ELECTRONIC_ADDRESSES);
-		
-		
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.ELECTRONIC_ADDRESSES);
+
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 		return msgDisplay;
 	}
@@ -877,29 +1005,87 @@ public class UpdateFacilityPage extends ViewFacilityPage {
 	 * @return					a string of the error message, if expectError is true. otherwise an empty string
 	 */
 	public String updateElectronicAddressDataBlock(String address, String effectiveFrom, String effectiveTo,
-												   int index, boolean expectError, String endReason) {
-		String msgDisplay="";
-		String formName=DIALOG_MAP.get(FacilitySection.ELECTRONIC_ADDRESSES).getFormName();
-		String dialogCss=getDialogCss(FacilitySection.ELECTRONIC_ADDRESSES);
+			EndReason endReasonCode, int index, boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.ELECTRONIC_ADDRESSES).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.ELECTRONIC_ADDRESSES);
 
 		clickDataBlockUpdateButton(FacilitySection.ELECTRONIC_ADDRESSES, index);
-		//waitSeconds(2);
+		waitSeconds(2);
 
-		String inputAddressCss=dialogCss+" >input#"+formName+"\\:"+"electronicAddress";
-		WebElement inputAddress=selenium_.findElement(By.cssSelector(inputAddressCss));
+		String inputAddressCss = dialogCss + " >input#" + formName + "\\:" + "electronicAddress";
+		WebElement inputAddress = selenium_.findElement(By.cssSelector(inputAddressCss));
 		inputAddress.clear();
-		if(!StringUtils.isEmpty(address))inputAddress.sendKeys(address);
+		if (!StringUtils.isEmpty(address))
+			inputAddress.sendKeys(address);
+		if (endReasonCode != null)
+			setEndReasonByVisibleText(FacilitySection.ELECTRONIC_ADDRESSES, endReasonCode.getText());
 
-		setEndReasonByVisibleText(FacilitySection.ELECTRONIC_ADDRESSES, endReason);
-
-		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ELECTRONIC_ADDRESSES,effectiveFrom,effectiveTo);
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ELECTRONIC_ADDRESSES, effectiveFrom, effectiveTo);
 
 		clickDialogSubmitButton(FacilitySection.ELECTRONIC_ADDRESSES);
 
-		if(expectError)
-			msgDisplay=waitErrorMessage(FacilitySection.ELECTRONIC_ADDRESSES);
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.ELECTRONIC_ADDRESSES);
 
 		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
 		return msgDisplay;
 	}
+	
+	/**
+	 * Click CHSA Button
+	 */
+	public void clickCHSAButton() {
+		if (grabDataBlockExpanded(FacilitySection.CIVIC_ADDRESSES, 0))
+			expandDataBlock(FacilitySection.CIVIC_ADDRESSES, 0, true);
+
+		WebElement updateChsaButton = selenium_
+				.waitUntil(ExpectedConditions.elementToBeClickable(By.xpath("//button[./span[text()='Update CHSA']]")));
+		selenium_.scrollIntoView(updateChsaButton);
+		updateChsaButton.click();
+
+		selenium_
+		.waitUntil(ExpectedConditions.stalenessOf(updateChsaButton));
+	}
+	
+	/**
+	 * Add Relationship By Auto-Completion
+	 *
+	 * @param org
+	 * @param relationType
+	 * @param effectiveFrom
+	 * @param effectiveTo
+	 * @param expectError
+	 * @return If a predicted exception occurs, it will capture and return the
+	 *         associated error message; otherwise, it waits for the dialog to
+	 *         dismiss before returning an empty string.
+	 */
+	public String addRelationshipByAutoCompletion(MaintainOrgBuilder org, String relationType, String effectiveFrom,
+			String effectiveTo, boolean expectError) {
+		String msgDisplay = "";
+		String formName = DIALOG_MAP.get(FacilitySection.ORGANIZATION_RELATIONSHIPS).getFormName();
+		String dialogCss = getDialogCss(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+
+		clickHeaderAddDateBlockButton(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+
+		String AUTOCOMPLETE_FIELD_CSS = "input#maintainOrgRelationshipForm\\:orgNameAutoComplete_input";
+		String panelCss = "span#maintainOrgRelationshipForm\\:orgNameAutoComplete";
+		AutocompleteMenu menu = new AutocompleteMenu(selenium_, By.cssSelector(AUTOCOMPLETE_FIELD_CSS),
+				By.cssSelector(panelCss));
+		menu.selectItemFromPanelMatch(org.getName(), org.getAddressList().get(0).get("line1"));
+
+		setDropdownListByVisibleText(FacilitySection.ORGANIZATION_RELATIONSHIPS, "relationshipType", relationType);
+
+		setDialogEffectiveFromAndEffectiveTo(FacilitySection.ORGANIZATION_RELATIONSHIPS, effectiveFrom, effectiveTo);
+
+		clickDialogSubmitButton(FacilitySection.ORGANIZATION_RELATIONSHIPS, expectError);
+
+		if (expectError)
+			msgDisplay = waitErrorMessage(FacilitySection.ORGANIZATION_RELATIONSHIPS);
+
+		selenium_.waitUntil(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(dialogCss)));
+		return msgDisplay;
+	}
+
+	
 }
